FUNCIONA: lista de expedientes en la lista de pacientes

+++ proyecto-fin-grado/src/main/java/com/example/proyectofingrado/service/impl/ExpendienteClinicoServiceImpl.java
@@ -26,7 +26,7 @@ public class ExpendienteClinicoServiceImpl implements ExpedienteClinicoService{
 
     @Override
     public void guardarExpedienteClinico(int idPaciente, ExpedienteClinicoDTO expedienteClinicoDTO) {
-        Paciente paciente = pacienteRepository.findById(idPaciente).orElseThrow(() -> new RuntimeException("Paciente no encontrado"));
+        Paciente paciente = pacienteRepository.findById((long) idPaciente).orElseThrow(() -> new RuntimeException("Paciente no encontrado"));
         expedienteClinicoDTO.setIdPaciente(paciente);
         ExpedienteClinico expedienteClinico = ExpedienteClinicoMapper.toExpedienteClinico(expedienteClinicoDTO);
         expedienteClinicoRepository.save(expedienteClinico);
@@ -34,16 +34,18 @@ public class ExpendienteClinicoServiceImpl implements ExpedienteClinicoService{
 
     @Override
     public List<ExpedienteClinicoDTO> obtenerExpedienteClinicosPorPaciente(int idPaciente) {
-        Paciente paciente = pacienteRepository.findById(idPaciente).orElseThrow(() -> new RuntimeException("Paciente no encontrado"));
+        Paciente paciente = pacienteRepository.findById((long) idPaciente)
+                .orElseThrow(() -> new RuntimeException("Paciente no encontrado"));
         List<ExpedienteClinico> expedienteClinicoList = expedienteClinicoRepository.findByPaciente(paciente);
         List<ExpedienteClinicoDTO> expedienteClinicoDTOList = expedienteClinicoList.stream().map((expedienteClinico) -> ExpedienteClinicoMapper.toExpedienteClinicoDTO(expedienteClinico))
                 .collect(Collectors.toList());
         return expedienteClinicoDTOList;
     }
 
+
     @Override
     public ExpedienteClinicoDTO getExpedienteClinicoById(int idExpedienteClinico) {
-        ExpedienteClinico expedienteClinico = expedienteClinicoRepository.findById(idExpedienteClinico).orElseThrow(() -> new RuntimeException("Expediente clínico no encontrado"));
+        ExpedienteClinico expedienteClinico = expedienteClinicoRepository.findById((long) idExpedienteClinico).orElseThrow(() -> new RuntimeException("Expediente clínico no encontrado"));
         ExpedienteClinicoDTO expedienteClinicoDTO = ExpedienteClinicoMapper.toExpedienteClinicoDTO(expedienteClinico);
         return expedienteClinicoDTO;
     }
@@ -55,6 +57,6 @@ public class ExpendienteClinicoServiceImpl implements ExpedienteClinicoService{
 
     @Override
     public void borrarExpedienteClinico(int idExpedienteClinico) {
-        expedienteClinicoRepository.deleteById(idExpedienteClinico);
+        expedienteClinicoRepository.deleteById((long) idExpedienteClinico);
     }
 }
